0.7.0: Playlists, Vorbis, ALAC and more
- XSPF playlists support
- Modernization of error handling and server-client communications
- Vorbis and ALAC transcoding
- All transcoders are now optional but enabled on default
- lonelyradio_types crate is now exported through monolib

+++ monoclient-x/monoclient-x/Metadata.swift
@@ -47,7 +47,7 @@ struct Cover {
 			self.cover = PlatformImage(cgImage: CGImage(jpegDataProviderSource: CGDataProvider(data: data)!, decode: nil, shouldInterpolate: false, intent: CGColorRenderingIntent.absoluteColorimetric)!).preparingForDisplay()!
 #endif
 			// deallocating memory
-			c_drop(cov.bytes, Int(cov.length))
+			c_drop(cov.bytes, UInt(Int(cov.length)))
 			print(self.cover.size)
 			
 		} else {
